// typechain-types/Verifier.ts
export interface VerifierInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "recoverSignerFromCompositeSignature"
      | "verifyCompositeSignature"
      | "verifyMessageInclusion"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "recoverSignerFromCompositeSignature",
    values: [BytesLike, BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "verifyCompositeSignature",
    values: [BytesLike, BytesLike[], BytesLike, BytesLike, AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "verifyMessageInclusion",
    values: [BytesLike, BytesLike[], BytesLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "recoverSignerFromCompositeSignature",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifyCompositeSignature",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifyMessageInclusion",
    data: BytesLike
  ): Result;
}

export interface Verifier extends BaseContract {
  connect(runner?: ContractRunner | null): Verifier;
  waitForDeployment(): Promise<this>;

  interface: VerifierInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  recoverSignerFromCompositeSignature: TypedContractMethod<
    [merkleRoot: BytesLike, signature: BytesLike],
    [string],
    "view"
  >;

  verifyCompositeSignature: TypedContractMethod<
    [
      messageHash: BytesLike,
      proof: BytesLike[],
      merkleRoot: BytesLike,
      signature: BytesLike,
      expectedSigner: AddressLike
    ],
    [boolean],
    "view"
  >;

  verifyMessageInclusion: TypedContractMethod<
    [messageHash: BytesLike, proof: BytesLike[], root: BytesLike],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "recoverSignerFromCompositeSignature"
  ): TypedContractMethod<
    [merkleRoot: BytesLike, signature: BytesLike],
    [string],
    "view"
  >;
  getFunction(
    nameOrSignature: "verifyCompositeSignature"
  ): TypedContractMethod<
    [
      messageHash: BytesLike,
      proof: BytesLike[],
      merkleRoot: BytesLike,
      signature: BytesLike,
      expectedSigner: AddressLike
    ],
    [boolean],
    "view"
  >;
  getFunction(
    nameOrSignature: "verifyMessageInclusion"
  ): TypedContractMethod<
    [messageHash: BytesLike, proof: BytesLike[], root: BytesLike],
    [boolean],
    "view"
  >;

  filters: {};
}